+++ bilrost-derive/src/field/ignored.rs
@@ -1,0 +1,46 @@
+use crate::attrs::word_attr;
+use alloc::vec;
+use alloc::vec::Vec;
+use eyre::{bail, Error};
+use proc_macro2::TokenStream;
+use quote::quote;
+use syn::{Meta, Type};
+
+#[derive(Clone)]
+pub(crate) struct Field {
+    ty: Type,
+}
+
+impl Field {
+    pub(crate) fn new(ty: &Type, attrs: &[Meta]) -> Result<Option<Self>, Error> {
+        let ignore_attr_count = attrs
+            .iter()
+            .filter(|attr| word_attr(attr, "ignore"))
+            .count();
+        if ignore_attr_count == 0 {
+            return Ok(None); // Field is not ignored
+        }
+        if ignore_attr_count > 1 {
+            bail!(
+                "duplicated ignore attrs for field: {}",
+                quote!(#(#attrs),*)
+            );
+        }
+        if attrs.len() > 1 {
+            bail!(
+                "ignore attribute mixed with other attrs on the same field: {}",
+                quote!(#(#attrs),*)
+            );
+        }
+        Ok(Some(Self { ty: ty.clone() }))
+    }
+
+    pub(crate) fn initialize(&self) -> TokenStream {
+        quote!(::core::default::Default::default())
+    }
+
+    pub(crate) fn where_terms(&self) -> Vec<TokenStream> {
+        let ty = &self.ty;
+        vec![quote!(#ty: ::core::default::Default)]
+    }
+}