+++ tests/derived_message_tests.rs
@@ -930,25 +930,21 @@ fn ignored_fields() {
 
 #[test]
 fn ignored_fields_with_defaults() {
+    #[derive(Debug, PartialEq)]
+    struct Undecodable(usize);
+
+    impl Default for Undecodable {
+        fn default() -> Self {
+            Undecodable(12345)
+        }
+    }
+
     #[derive(Debug, PartialEq, Message)]
     struct FooPlus {
         x: i64,
         y: i64,
         #[bilrost(ignore)]
-        also: usize,
-    }
-
-    // Some Default implementation is required when there are ignored fields. It doesn't have
-    // to be the derived implementation, and it can have non-empty values for non-ignored
-    // fields.
-    impl Default for FooPlus {
-        fn default() -> Self {
-            Self {
-                x: 111,
-                y: 222,
-                also: 12345,
-            }
-        }
+        also: Undecodable,
     }
 
     // The empty value for the message will still have the empty value for all non-ignored
@@ -958,7 +954,7 @@ fn ignored_fields_with_defaults() {
         FooPlus {
             x: 0,
             y: 0,
-            also: 12345,
+            also: Undecodable(12345),
         }
     );
 
@@ -968,7 +964,7 @@ fn ignored_fields_with_defaults() {
         FooPlus {
             x: 1,
             y: 0,
-            also: 12345,
+            also: Undecodable(12345),
         },
     )
 }
@@ -1189,15 +1185,6 @@ fn generic_encodings() {
     #[allow(dead_code)]
     #[derive(Message)]
     struct Foo<T: Default, E>(#[bilrost(encoding(E))] T, #[bilrost(ignore)] PhantomData<E>);
-
-    impl<T, E> Default for Foo<T, E>
-    where
-        T: Default,
-    {
-        fn default() -> Self {
-            Self(Default::default(), PhantomData)
-        }
-    }
 
     static_assertions::assert_impl_all!(Foo<String, General>: OwnedMessage);
     static_assertions::assert_not_impl_any!(Foo<u8, General>: Message);